Add support for setting the MAC address
This adds a `:mac_address` parameter that can be set to either a string
or MFA tuple. When specified, VintageNet will set the MAC address on the
network interface before bringing the interface up.

Here's an example configuration:

```elixir
{"eth0",
  %{
    type: VintageNetEthernet,
    mac_address: {MyMacAddressReader, :read, []},
    ipv4: %{method: :dhcp}
  }}
```

In the example, VintageNet calls `MyMacaddressReader.read()` for the MAC
address.